requirements.txt
openai==2.26.0
faiss-cpu==1.13.2
numpy==2.4.2
beautifulsoup4==4.14.3
requests==2.32.5
python-dotenv==1.2.2
